+++ demo/src/image-builder/commands/actions/adjust/index.ts
@@ -8,7 +8,12 @@ import {
   Vec2,
   Vector2,
   EntityObject,
-  MathUtils
+  MathUtils,
+  CommandManager,
+  HotKey,
+  Key,
+  HotKeyEventType,
+  InferenceEngine
 } from '@turbox3d/turbox';
 
 import { ItemEntity } from '../../../models/entity/item';
@@ -19,23 +24,47 @@ const ACTION_NAME = 'adjustEntity';
 export class AdjustCommand extends Command {
   private adjustAction = Action.create(ACTION_NAME);
   private target?: ItemEntity;
+  private inferenceEngine = new InferenceEngine();
+  private fixedScale = false;
+
+  constructor(holder: CommandManager) {
+    super(holder);
+    HotKey.on({
+      key: Key.Ctrl,
+      handler: (keyEventType: HotKeyEventType) => {
+        const selected = appCommandManager.defaultCommand.select.getSelectedEntities();
+        if (selected.length > 0) {
+          this.fixedScale = true;
+        }
+        if (keyEventType === HotKeyEventType.KeyUp) {
+          this.fixedScale = false;
+        }
+      },
+    })
+  }
 
   onAdjustStartHandler(v: Partial<ViewEntity>, e: SceneEvent, t: SceneTool) {
     const selected = appCommandManager.defaultCommand.select.getSelectedEntities()[0];
     if (selected instanceof ItemEntity && !selected.locked) {
       this.target = selected;
-      this.onRotateStartHandler(v, e, t);
+      if (!this.fixedScale) {
+        this.onRotateStartHandler(v, e, t);
+      }
       this.onScaleStartHandler(v, e, t);
     }
   }
 
   onAdjustMoveHandler(v: Partial<ViewEntity>, e: SceneEvent, t: SceneTool) {
-    this.onRotateMoveHandler(v, e, t);
+    if (!this.fixedScale) {
+      this.onRotateMoveHandler(v, e, t);
+    }
     this.onScaleMoveHandler(v, e, t);
   }
 
   onAdjustEndHandler(v: Partial<ViewEntity>, e: SceneEvent, t: SceneTool) {
-    this.onRotateEndHandler(v, e, t);
+    if (!this.fixedScale) {
+      this.onRotateEndHandler(v, e, t);
+    }
     this.onScaleEndHandler(v, e, t);
   }
 
@@ -60,7 +89,7 @@ export class AdjustCommand extends Command {
     const localPoint = new Vector2(mp.x, mp.y).applyMatrix3(this.target.getConcatenatedMatrix3(type).inverted());
     const degree = this.initPosition.clone().angleTo(localPoint) * MathUtils.RAD2DEG;
     this.degree = degree;
-    const { snapped } = this.snap(this.target.rotation.z + degree);
+    const { snapped } = this.inferenceEngine.rotateSnap(this.target.rotation.z + degree);
     this.adjustAction.execute(
       () => {
         this.target?.setRotation({
@@ -81,11 +110,11 @@ export class AdjustCommand extends Command {
       this.adjustAction.abort();
       this.adjustAction = Action.create(ACTION_NAME);
     } else {
-      const { rotationZ } = this.snap(this.target.rotation.z + this.degree);
+      const { snappedDegree } = this.inferenceEngine.rotateSnap(this.target.rotation.z + this.degree);
       this.adjustAction.execute(
         () => {
           this.target?.setRotation({
-            z: rotationZ,
+            z: snappedDegree,
           });
           this.target?.$update({
             snapped: false,
@@ -103,27 +132,10 @@ export class AdjustCommand extends Command {
     this.degree = 0;
   }
 
-  private snap(degree: number) {
-    let rotationZ = degree % 360;
-    let snapped = false;
-    const baseLine = 90;
-    const snapDegree = 15;
-    if (Math.abs(rotationZ % baseLine) <= snapDegree) {
-      rotationZ = Math.floor(rotationZ / baseLine) * baseLine;
-      snapped = true;
-    } else if (Math.abs(rotationZ % baseLine) >= baseLine - snapDegree) {
-      rotationZ = Math.ceil(rotationZ / baseLine) * baseLine;
-      snapped = true;
-    }
-    return {
-      rotationZ,
-      snapped,
-    };
-  }
-
   private initSize?: Vec2;
   private initLength?: number;
   private initFontSize?: number;
+  private initItemPosition?: Vec2;
 
   private onScaleStartHandler(viewEntity: Partial<ViewEntity>, event: SceneEvent, tools: SceneTool) {
     if (!this.target) {
@@ -138,6 +150,10 @@ export class AdjustCommand extends Command {
       y: this.target.size.y,
     };
     this.initFontSize = this.target.fontSize;
+    this.initItemPosition = {
+      x: this.target.position.x,
+      y: this.target.position.y,
+    };
   }
 
   private onScaleMoveHandler(viewEntity: Partial<ViewEntity>, event: SceneEvent, tools: SceneTool) {
@@ -148,19 +164,37 @@ export class AdjustCommand extends Command {
     const mp = event.getScenePosition() as Vec2;
     const type = EntityObject.EPerspectiveType.FRONT;
     const localPoint = new Vector2(mp.x, mp.y).applyMatrix3(itemEntity.getConcatenatedMatrix3(type).inverted());
-    if (!this.initLength || !this.initSize || !this.initFontSize) {
+    if (!this.initLength || !this.initSize || !this.initFontSize || !this.initItemPosition) {
       return;
     }
     this.adjustAction.execute(
       () => {
-        const scale = localPoint.length / this.initLength!;
-        itemEntity.setSize({
-          x: this.initSize!.x * scale,
-          y: this.initSize!.y * scale,
-        });
-        itemEntity.$update({
-          fontSize: this.initFontSize! * scale,
-        });
+        if (this.fixedScale) {
+          const sub = new Vector2(mp.x, mp.y).subtracted(new Vector2(this.initItemPosition!.x, this.initItemPosition!.y));
+          const scale = (sub.length / this.initLength!) * (sub.x >= 0 ? 1 : -1);
+          if (scale > -1) {
+            itemEntity.setPosition({
+              x: this.initItemPosition!.x + (this.initSize!.x / 2 * scale + this.initSize!.x / 2) / 2 - this.initSize!.x / 2,
+              y: this.initItemPosition!.y + (this.initSize!.y / 2 * scale + this.initSize!.y / 2) / 2 - this.initSize!.y / 2,
+            });
+            itemEntity.setSize({
+              x: this.initSize!.x / 2 * scale + this.initSize!.x / 2,
+              y: this.initSize!.y / 2 * scale + this.initSize!.y / 2,
+            });
+            itemEntity.$update({
+              fontSize: this.initFontSize! / 2 * scale + this.initFontSize! / 2,
+            });
+          }
+        } else {
+          const scale = localPoint.length / this.initLength!;
+          itemEntity.setSize({
+            x: this.initSize!.x * scale,
+            y: this.initSize!.y * scale,
+          });
+          itemEntity.$update({
+            fontSize: this.initFontSize! * scale,
+          });
+        }
       },
       undefined,
       true,
@@ -180,5 +214,6 @@ export class AdjustCommand extends Command {
     this.initLength = undefined;
     this.initSize = undefined;
     this.initFontSize = undefined;
+    this.initItemPosition = undefined;
   }
 }