+++ packages/command-manager/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @turbox3d/command-manager
 
+## 1.1.28
+
+### Patch Changes
+
+- fix: fix
+- Updated dependencies
+  - @turbox3d/event-manager@1.1.27
+  - @turbox3d/shared@1.1.20
+
 ## 1.1.27
 
 ### Patch Changes