+++ packages/design-engine/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @turbox3d/design-engine
 
+## 1.1.32
+
+### Patch Changes
+
+- fix: fix
+- Updated dependencies
+  - @turbox3d/command-manager@1.1.28
+  - @turbox3d/event-manager@1.1.27
+  - @turbox3d/math@1.1.17
+  - @turbox3d/reactivity@2.1.24
+  - @turbox3d/reactivity-react@1.1.24
+  - @turbox3d/shared@1.1.20
+
 ## 1.1.31
 
 ### Patch Changes