+++ packages/reactivity-react/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @turbox3d/reactivity-react
 
+## 1.1.24
+
+### Patch Changes
+
+- fix: fix
+- Updated dependencies
+  - @turbox3d/reactivity@2.1.24
+  - @turbox3d/shared@1.1.20
+
 ## 1.1.23
 
 ### Patch Changes